requirements.txt
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
Flask==3.0.3
html5lib==1.1
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kiwisolver==1.3.2
Mako==1.1.3
MarkupSafe==3.0.2
matplotlib==3.5.1
MAVProxy==1.8.71
monotonic==1.6
more-itertools==8.10.0
numpy==2.1.2
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
pygame==2.1.2
pymacaroons==0.13.0
pymavlink==2.4.41
PyNaCl==1.5.0
pynmeagps==1.0.43
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
python-dateutil==2.8.1
pytz==2022.1
PyVISA==1.14.1
PyVISA-py==0.7.2
pyxdg==0.27
PyYAML==6.0.2
reportlab==3.6.8
requests==2.25.1
rplidar==0.9.2
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
ssh-import-id==5.11
urllib3==1.26.5
pyvisa-py
